Commit on 15-10-2025
Fixed rental request emails — now two emails are sent correctly if there are two products from the same company.
Updated email templates with new header and footer.
Added email notifications for when a user sends an offer to a supplier.
Created a new API to update job names in the supply jobs list.